moved privilege check to group level from api layer to service layer
